license for shared

+++ util/misc.sh
@@ -1,3 +1,56 @@
+experiments() {
+    for f in ${*:-`pwd`}; do
+        experimentf $f/*/my.experiment.yml $f/my.experiment.yml
+    done
+}
+tunes() {
+    for f in ${*:-`pwd`}; do
+        ag -G "iter_${iter}"'.*\.err.*'$parti 'Converged' $f
+    done
+}
+expclean() {
+    for f in ${*:-`pwd`}; do
+        expcleanf $f/*/my.experiment.yml $f/my.experiment.yml
+    done
+}
+expcleanf() {
+    for f in "$@"; do
+        if [[ -f $f ]] ; then
+            local dir=`dirname $f`
+            rm -rf $dir/tune_slot/tune_work
+        fi
+    done
+}
+experimentf() {
+    for f in "$@"; do
+        if [[ -f $f ]] ; then
+            local dir=`dirname $f`
+            perl -e 'while(<>) { print "$2\t" if /([a-zA-Z_-]+)bleu: '"'?-?([0-9.]+)/ }" $f
+            perl -e '$n = 0; $z = 0;while(<>) { chomp;
+                      ++$n if '"/: '?-?[0-9.]/"'; ++$z if '"/: '?-?0+\.?0*'?$/"';
+                     } print "$n\t$z\t"' $dir/config/weights.file.final
+            echo $dir
+        fi
+    done
+}
+smert() {
+    save12 ~/tmp/cmert cwithmertargs -x -f 0 /home/graehl/projects/sparse/weights /home/graehl/projects/sparse/${nbest:-nbest.10k} "$@"
+}
+cmert() {
+    save12 ~/tmp/cmert  cwithmertargs  -x -f 0 /home/graehl/projects/tune/tune_work/iter_0/initial.txt.19 /home/graehl/projects/tune/tune_work/iter_0/output.nbest/corpus.nbest "$@"
+    if [[ $l0bonus ]] ; then
+        c-s "cp -a /home/graehl/c/ct-archive/archive/3rdParty/mert/mert /home/graehl/pub/mert-l0=${l0bonus}; ls ~/pub/mert*"
+    fi
+    if [[ $CC ]] ; then
+        c-s "set -x; cp -a /home/graehl/c/ct-archive/archive/3rdParty/mert/mert /home/graehl/pub/mert-$CC"
+    fi
+}
+l0s="0 1e-2 4e-2 1e-3 4e-3 4e-4 1e-5 4e-5 4e-6 1e-4"
+cmerts() {
+    for f in $l0s; do
+        SAN= l0bonus=$f CC=gcc DEBUG= ASSERT= cmert
+    done
+}
 benchmert() {
 for j in 6 8 10 12 14 16 18 32 64; do
 echo2 ===== j=$j
@@ -14,11 +67,11 @@ theirbaseline() {
 }
 initweights() {
     set -e
-    local s=$1/config/weights.file.final
-    local d=$2/config/weights.file.final
+    local d=$1/config/weights.file.final
+    local s=$2/config/weights.file.final
     require_files $s $d
-    cp $s init.sparse.yml
-    cp $d init.dense.yml
+    [ -f init.sparse.yml ] || cp $s init.sparse.yml
+    [ -f init.dense.yml ] || cp $d init.dense.yml
     wc -l init.sparse.yml init.dense.yml
 }
 baselinetune() {
@@ -44,7 +97,6 @@ mycommands() {
     require_file $xmt
     xmt=`abspath $xmt`
     require_file $theirs $tunestart
-    l0s="0 4e-6 1e-5 4e-5 4e-4 1e-4"
     for sd in $densesparse; do
         origweights=$lp/init.$sd.yml
         require_file $origweights
@@ -99,7 +151,7 @@ lntune() {
     require_dir $lp
     theirs=$lp/their.apex
     cp $fromapex $theirs
-    mine=$lp/my.apex
+    mine=$lp/my.$densesparse.apex
 }
 mytune() {
     set -e
